+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/extensions/SentryPluginExtension.kt
@@ -105,6 +105,19 @@ abstract class SentryPluginExtension @Inject constructor(project: Project) {
     autoInstallationAction.execute(autoInstallation)
   }
 
+  val sizeAnalysis: SizeAnalysisExtension = objects.newInstance(SizeAnalysisExtension::class.java)
+
+  fun sizeAnalysis(sizeAnalysisAction: Action<SizeAnalysisExtension>) {
+    sizeAnalysisAction.execute(sizeAnalysis)
+  }
+
+  val vcsInfo: VcsInfoExtension = objects.newInstance(VcsInfoExtension::class.java)
+
+  /** Configure the VCS information for build uploads. */
+  fun vcsInfo(vcsInfoAction: Action<VcsInfoExtension>) {
+    vcsInfoAction.execute(vcsInfo)
+  }
+
   /**
    * Disables or enables the reporting of dependencies metadata for Sentry. If enabled the plugin
    * will collect external dependencies and will take care of uploading them to Sentry as part of